# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Goalsy Executive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List in-app notifications (non-dismissed)
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: clearNotifications
      tags: [notifications]
      summary: Dismiss all notifications
      responses:
        "204":
          description: Cleared
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/{id}/dismiss:
    post:
      operationId: dismissNotification
      tags: [notifications]
      summary: Dismiss a notification
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: ["string", "null"]
        avatarKey:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [userId, onboardingComplete, createdAt, updatedAt]
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        avatarKey:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
...
    AppNotification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        targetScreen:
          type: ["string", "null"]
        targetId:
          type: ["string", "null"]
        isRead:
          type: boolean
        isDismissed:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, type, title, isRead, isDismissed, createdAt]